requirements.txt

pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
mlflow==2.3.1
fastapi==0.95.2
uvicorn==0.22.0
pyyaml==6.0
joblib==1.2.0
